Docs: Added docs for databse unit
Refactor: Moved soci custom types to new private-impl file
